requirements.txt
openai
python-dotenv
tabulate
sounddevice
numpy
pandas
langchain
langchain-openai
langchain-experimental
markdown
django-widget-tweaks
chromadb
langchain
langchain_openai
pypdf
